chore(frontend): pulls event sending and handling logic out into an events.ts file
